requirement.txt
django==1.10.3
MySQL-python==1.2.5
Pillow==4.0.0
